+++ src/components/mantle/button.tsx
@@ -2,9 +2,9 @@ import { type ButtonHTMLAttributes, forwardRef } from "react";
 import { Slot } from "@radix-ui/react-slot";
 import { cva } from "class-variance-authority";
 
-import { cx } from "../lib/cx";
-import type { WithAsChild } from "../types/as-child";
-import type { VariantProps } from "../types/variant-props";
+import { cx } from "../../lib/cx";
+import type { WithAsChild } from "../../types/as-child";
+import type { VariantProps } from "../../types/variant-props";
 
 const buttonVariants = cva(
 	"inline-flex items-center justify-center rounded-md font-medium border border-transparent transition-colors focus-visible:outline-none focus-visible:ring-4 disabled:pointer-events-none disabled:opacity-50",
