+++ src/components/mantle/media-object.tsx
@@ -1,7 +1,7 @@
 import { forwardRef } from "react";
 import type { HTMLAttributes } from "react";
 
-import { cx } from "../lib/cx";
+import { cx } from "../../lib/cx";
 
 type Props = HTMLAttributes<HTMLDivElement>;
