+++ src/layouts/PageLayout.astro
@@ -1,7 +1,7 @@
 ---
 import "../mantle.css";
-import { Input } from "../components/input";
-import { Button } from "../components/button";
+import { Input } from "../components/mantle/input";
+import { Button } from "../components/mantle/button";
 
 const { title } = Astro.props;
 ---
